+++ Changelog.md
@@ -11,6 +11,8 @@ Bug Fixes:
   (Jon Rowe, Phil Pirozhkov, rspec/rspec-rails#2909)
 * Fix view specs manipulating a frozen array (for template prefixes) in `actionview` internals in Rails 8.2.
   (Iliana Hadzhiatanasova, rspec/rspec-rails#2915)
+* Keep controller specs from rendering views reached through a view path added as a string.
+  (Phil Pirozhkov, rspec/rspec-rails#2917)
 
 ### 8.0.4 / 2026-03-10
 [Full Changelog](https://github.com/rspec/rspec-rails/compare/v8.0.3...v8.0.4)